+++ src/domains/store/store.service.ts
@@ -4,8 +4,8 @@ import type { ProductWithStock } from './store.mapper.js';
 import { ConflictError, NotFoundError, ForbiddenError } from '@/common/utils/errors.js';
 import { logger } from '@/config/logger.js';
 import { env } from '@/config/constants.js';
-import { SecurityEventType } from '@/common/types/security-events.type.js';
-import { sanitizePhoneNumber, measureDuration } from '@/common/utils/logger-helpers.js';
+import { SecurityEventType } from '@/common/types/securityEvents.type.js';
+import { sanitizePhoneNumber, measureDuration } from '@/common/utils/loggerHelpers.js';
 
 export class StoreService {
   constructor(private storeRepository: StoreRepository) {}
